requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
PyMySQL==1.1.1
python-dotenv==1.0.1
pydantic==2.9.1
ruff==0.6.3
